Separated structs from GooglSearchService to make into reusable Models
Following MVVM foundations and best practices, the following files were made: Article.swift, GoogleSearchItem.swift and GoogleSearchResponse.swift.
These were made into models that are also reusable.
GoogleSearchService.swift: removed the structs that are now Models.
